useMediaAutoDownload working for images
