+++ app/i18n/locales/en.json
@@ -317,7 +317,7 @@
   "How_It_Works": "How it works",
   "I_Saved_My_E2E_Password": "I saved my E2E password",
   "Ignore": "Ignore",
-  "Images": "Images",
+  "Image": "Image",
   "Images_uploaded": "Images uploaded",
   "In_app": "In-app",
   "In_App_And_Desktop": "In-app and desktop",
