+++ app/lib/constants/mediaAutoDownload.ts
@@ -1,5 +1,5 @@
 export type MediaDownloadOption = 'never' | 'wifi_mobile_data' | 'wifi';
 
-export const IMAGES_PREFERENCE_DOWNLOAD = 'imagesPreferenceDownload';
+export const IMAGE_PREFERENCE_DOWNLOAD = 'imagePreferenceDownload';
 export const VIDEO_PREFERENCE_DOWNLOAD = 'videoPreferenceDownload';
 export const AUDIO_PREFERENCE_DOWNLOAD = 'audioPreferenceDownload';
